requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.2.0
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
dnspython==2.7.0
docker==7.1.0
email_validator==2.1.2
fastapi==0.115.0
fastapi-cli==0.0.5
fastapi-users==12.1.3
fastapi-users-db-sqlalchemy==6.0.1
greenlet==3.1.1
h11==0.12.0
httpcore==0.13.3
httptools==0.6.1
httpx==1.0.0b0
idna==3.10
inflect==7.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
makefun==1.15.6
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==3.0.0
mdurl==0.1.2
more-itertools==10.5.0
mypy-extensions==1.0.0
orjson==3.10.7
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.9
pwdlib==0.2.1
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.7
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
rfc3986==1.5.0
rich==13.9.2
setuptools==75.1.0
shellingham==1.5.4
sniffio==1.3.1
sqlacodegen==3.0.0rc5
SQLAlchemy==2.0.35
starlette==0.38.6
typeguard==4.3.0
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.31.0
uvloop==0.21.0b1
watchfiles==0.24.0
websockets==13.1
wheel==0.44.0
